pdf_and_doc_generation/requirements.txt
fpdf
python-docx
